Rename GoogleMapsMapProvider to MapProvider
 * All gms specific code is moved to the GoogleMapsFragment
 * Remaining will be moved in the followup commit
