--- gnd/src/main/java/com/google/android/gnd/ui/map/MapProviderModule.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.gnd.ui.map;
-
-import com.google.android.gnd.ui.map.gms.GoogleMapsMapProvider;
-import dagger.Binds;
-import dagger.Module;
-import dagger.hilt.InstallIn;
-import dagger.hilt.components.SingletonComponent;
-import javax.inject.Singleton;
-
-@InstallIn(SingletonComponent.class)
-@Module
-public abstract class MapProviderModule {
-
-  @Binds
-  @Singleton
-  abstract MapProvider googleMapsProvider(GoogleMapsMapProvider mapsMapProvider);
-}
